Added applType config + Screenshots
First of all, I removed the thread testing folder since we do not need
the uneccesary clutter.

Secondly, I added an application type to the omnetpp.ini file so there
is no longer a prompt to specify the applType template.

Finally, I've taken screenshots of the description tab for each of the
.ned files so that I can add a graphical representation of each NED to
the readme.md.